move filtering function to misc
